+++ api/connection.js
@@ -3,7 +3,7 @@ const mysql = require('mysql2');
 const connection = mysql.createConnection({
     host: 'localhost',
     user: 'root',
-    password: '',
+    password: 'easi',
     database: 'easeengenharia',
     multipleStatements: true
 });
