+++ api/db/script.sql
@@ -12,25 +12,6 @@ CREATE TABLE `user` (
    UNIQUE KEY `email` (`email`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
-DROP TABLE IF EXISTS `item`;
-CREATE TABLE `item` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `title` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
-  `observation` text COLLATE utf8_unicode_ci,
-  `rating` tinyint(4) DEFAULT '0',
-  `image` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
-  `created` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-  PRIMARY KEY (`id`) USING BTREE
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
-
-DROP TABLE IF EXISTS `floor`;
-CREATE TABLE `floor` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `title` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
-  `created` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-  PRIMARY KEY (`id`) USING BTREE
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
-
 DROP TABLE IF EXISTS `construction`;
 CREATE TABLE `construction` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
@@ -40,42 +21,49 @@ CREATE TABLE `construction` (
   PRIMARY KEY (`id`) USING BTREE
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
 
-DROP TABLE IF EXISTS `construction_floor`;
-CREATE TABLE `construction_floor` (
-  `id_construction` int(11) NOT NULL,
-  `id_floor` int(11) NOT NULL,
-  PRIMARY KEY (`id_construction`, `id_floor`) USING BTREE,
-  CONSTRAINT `fk_cf_construction` FOREIGN KEY (`id_construction`) REFERENCES `construction` (`id`) ON DELETE CASCADE,
-  CONSTRAINT `fk_cf_floor` FOREIGN KEY (`id_floor`) REFERENCES `floor` (`id`) ON DELETE CASCADE
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
-
-DROP TABLE IF EXISTS `floor_item`;
-CREATE TABLE `floor_item` (
-  `id_floor` int(11) NOT NULL,
-  `id_item` int(11) NOT NULL,
-  PRIMARY KEY (`id_floor`, `id_item`) USING BTREE,
-  CONSTRAINT `fk_fi_item` FOREIGN KEY (`id_item`) REFERENCES `item` (`id`) ON DELETE CASCADE,
-  CONSTRAINT `fk_fi_floor` FOREIGN KEY (`id_floor`) REFERENCES `floor` (`id`) ON DELETE CASCADE
-) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
-
 DROP TABLE IF EXISTS `report`;
 CREATE TABLE `report` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id_construction` int(11) NOT NULL,
   `service_number` int(11) DEFAULT 0,
   `pdf` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
   `word` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
   `created` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-  PRIMARY KEY (`id`) USING BTREE
+  PRIMARY KEY (`id`) USING BTREE,
+  CONSTRAINT `fk_r_construction` FOREIGN KEY (`id_construction`) REFERENCES `construction` (`id`) ON DELETE CASCADE
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
 
-DROP TABLE IF EXISTS `construction_report`;
-CREATE TABLE `construction_report` (
+DROP TABLE IF EXISTS `floor`;
+CREATE TABLE `floor` (
+  `id` int(11) NOT NULL AUTO_INCREMENT,
   `id_construction` int(11) NOT NULL,
-  `id_report` int(11) NOT NULL,
-  PRIMARY KEY (`id_construction`, `id_report`) USING BTREE,
-  CONSTRAINT `fk_cr_construction` FOREIGN KEY (`id_construction`) REFERENCES `construction` (`id`) ON DELETE CASCADE,
-  CONSTRAINT `fk_cr_report` FOREIGN KEY (`id_report`) REFERENCES `report` (`id`) ON DELETE CASCADE
+  `title` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
+  `created` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
+  PRIMARY KEY (`id`) USING BTREE,
+  CONSTRAINT `fk_f_construction` FOREIGN KEY (`id_construction`) REFERENCES `construction` (`id`) ON DELETE CASCADE
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+
+DROP TABLE IF EXISTS `item`;
+CREATE TABLE `item` (
+  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id_floor` int(11) NOT NULL,
+  `title` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
+  `observation` text COLLATE utf8_unicode_ci,
+  `rating` tinyint(4) DEFAULT '0',
+  `image` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
+  `created` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
+  PRIMARY KEY (`id`) USING BTREE,
+  CONSTRAINT `fk_i_floor` FOREIGN KEY (`id_floor`) REFERENCES `floor` (`id`) ON DELETE CASCADE
+) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
+
+-- DROP TABLE IF EXISTS `construction_floor`;
+-- CREATE TABLE `construction_floor` (
+--   `id_construction` int(11) NOT NULL,
+--   `id_floor` int(11) NOT NULL,
+--   PRIMARY KEY (`id_construction`, `id_floor`) USING BTREE,
+--   CONSTRAINT `fk_cf_construction` FOREIGN KEY (`id_construction`) REFERENCES `construction` (`id`) ON DELETE CASCADE,
+--   CONSTRAINT `fk_cf_floor` FOREIGN KEY (`id_floor`) REFERENCES `floor` (`id`) ON DELETE CASCADE
+-- ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
 
 CREATE PROCEDURE `user_add_or_edit` (
 	IN _id INT,
